+++ src/optimum/rbln/transformers/models/decoderonly/__init__.py
@@ -12,18 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ....ops import (
-    custom_moe_ff,
-    custom_moe_glu,
-    paged_attn_decode,
-    paged_attn_prefill,
-    paged_causal_attn_decode,
-    paged_causal_attn_prefill,
-    paged_flash_attn_decode,
-    paged_flash_attn_prefill,
-    paged_flash_causal_attn_decode,
-    paged_flash_causal_attn_prefill,
-)
 from .configuration_decoderonly import RBLNDecoderOnlyModelConfig, RBLNDecoderOnlyModelForCausalLMConfig
 from .configuration_lora import RBLNLoRAAdapterConfig, RBLNLoRAConfig
 from .modeling_decoderonly import RBLNDecoderOnlyModel, RBLNDecoderOnlyModelForCausalLM
