+++ src/optimum/rbln/transformers/models/gpt_oss/gpt_oss_architecture.py
@@ -19,7 +19,7 @@ import torch
 import torch.nn.functional as F
 from torch import nn
 
-from ....ops.moe import compute_masked_routing_weight_topk_first
+from ...utils.moe import compute_masked_routing_weight_topk_first
 from ..decoderonly.configuration_decoderonly import RBLNLoRAConfig
 from ..decoderonly.decoderonly_architecture import (
     DecoderOnlyAttention,
