+++ src/optimum/rbln/transformers/models/mixtral/mixtral_architecture.py
@@ -17,7 +17,7 @@ from typing import Optional
 import torch
 from torch import nn
 
-from ....ops.moe import compute_masked_routing_weight_softmax_first
+from ...utils.moe import compute_masked_routing_weight_softmax_first
 from ..decoderonly.configuration_decoderonly import RBLNLoRAConfig
 from ..decoderonly.decoderonly_architecture import DecoderOnlyAttention, DecoderOnlyLayer, DecoderOnlyWrapper
