+++ src/optimum/rbln/transformers/models/t5/__init__.py
@@ -12,6 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ....ops import paged_add_softmax_attn_decode
 from .configuration_t5 import RBLNT5EncoderModelConfig, RBLNT5ForConditionalGenerationConfig
 from .modeling_t5 import RBLNT5EncoderModel, RBLNT5ForConditionalGeneration
