+++ src/optimum/rbln/transformers/models/time_series_transformer/__init__.py
@@ -21,6 +21,5 @@
 # copied, modified, or distributed without prior written permission
 # from Rebellions Inc.
 
-from ....ops import paged_add_softmax_attn_decode, rbln_cache_update
 from .configuration_time_series_transformer import RBLNTimeSeriesTransformerForPredictionConfig
 from .modeling_time_series_transformer import RBLNTimeSeriesTransformerForPrediction
