+++ src/optimum/rbln/transformers/models/whisper/__init__.py
@@ -12,6 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ....ops import paged_add_softmax_attn_decode
 from .configuration_whisper import RBLNWhisperForConditionalGenerationConfig
 from .modeling_whisper import RBLNWhisperForConditionalGeneration
